move imports to within method call
